renamed project name for android and overall
