+++ android/app/src/main/java/com/awesomeproject/MainActivity.java
@@ -1,4 +1,4 @@
-package com.huge;
+package com.awesomeproject;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "HUGE";
+    return "AwesomeProject";
   }
 }
