+++ android/app/src/main/java/com/awesomeproject/MainApplication.java
@@ -1,4 +1,4 @@
-package com.huge;
+package com.awesomeproject;
 
 import android.app.Application;
 import android.content.Context;
