+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "HUGE",
-  "displayName": "HUGE"
+  "name": "AwesomeProject",
+  "displayName": "AwesomeProject"
 }
