+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "HUGE",
+  "name": "AwesomeProject",
   "version": "0.0.1",
   "private": true,
   "scripts": {
